[GTK][WPE] Rename AcceleratedBackingStoreDMABuf as AcceleratedBackingStore https://bugs.webkit.org/show_bug.cgi?id=296477
Reviewed by Adrian Perez de Castro.

It can also handle shared memory buffers and in the case of WPE more buffer types will be added soon.

Canonical link: https://commits.webkit.org/297914@main
